# server/src/generated/shogi-ai/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: `${string}://openapi.json` | (string & {});
};

/**
 * AnalysisRequest
 *
 * 局面解析リクエスト
 */
export type AnalysisRequest = {
    /**
     * Sfen
     *
     * SFEN形式の局面文字列
     */
    sfen?: string | null;
    /**
     * Moves
     *
     * 初期局面からの指し手リスト（USI形式）
     */
    moves?: Array<string> | null;
    /**
     * Time Ms
     *
     * 思考時間（ミリ秒）
     */
    time_ms?: number;
    /**
     * Depth
     *
     * 探索深さ（指定時はtime_msを無視）
     */
    depth?: number | null;
    /**
     * Multipv
     *
     * 候補手の数（MultiPV）
     */
    multipv?: number;
};

/**
 * AnalysisResponse
 *
 * 局面解析レスポンス
 */
export type AnalysisResponse = {
    /**
     * Bestmove
     *
     * 最善手（USI形式）
     */
    bestmove: string;
    /**
     * Variations
     *
     * 候補手リスト（MultiPV）
     */
    variations: Array<MoveInfo>;
    /**
     * Time Ms
     *
     * 実際の思考時間
     */
    time_ms: number;
    /**
     * Engine Name
     *
     * エンジン名
     */
    engine_name: string;
};

/**
 * EngineInfoResponse
 *
 * エンジン情報レスポンス
 */
export type EngineInfoResponse = {
    /**
     * Name
     */
    name: string;
    /**
     * Author
     */
    author: string;
    /**
     * Version
     */
    version: string;
    /**
     * Options
     */
    options: {
        [key: string]: string;
    };
};

/**
 * HTTPValidationError
 */
export type HttpValidationError = {
    /**
     * Detail
     */
    detail?: Array<ValidationError>;
};

/**
 * MateSearchRequest
 *
 * 詰み探索リクエスト
 */
export type MateSearchRequest = {
    /**
     * Sfen
     *
     * SFEN形式の局面文字列
     */
    sfen?: string | null;
    /**
     * Moves
     *
     * 初期局面からの指し手リスト
     */
    moves?: Array<string> | null;
    /**
     * Time Ms
     *
     * 詰み探索時間（ミリ秒）
     */
    time_ms?: number;
};

/**
 * MateSearchResponse
 *
 * 詰み探索レスポンス
 */
export type MateSearchResponse = {
    /**
     * Mate Found
     *
     * 詰みが見つかったか
     */
    mate_found: boolean;
    /**
     * Mate Moves
     *
     * 詰み手順（USI形式）
     */
    mate_moves?: Array<string> | null;
    /**
     * Mate Length
     *
     * 詰みまでの手数
     */
    mate_length?: number | null;
    /**
     * Message
     *
     * メッセージ
     */
    message?: string | null;
};

/**
 * MoveInfo
 *
 * 1つの候補手情報
 */
export type MoveInfo = {
    /**
     * Move
     *
     * 指し手（USI形式）
     */
    move: string;
    /**
     * Score Cp
     *
     * 評価値（センチポーン）
     */
    score_cp?: number | null;
    /**
     * Score Mate
     *
     * 詰みまでの手数（プライ数）
     */
    score_mate?: number | null;
    /**
     * Depth
     *
     * 探索深さ
     */
    depth: number;
    /**
     * Nodes
     *
     * 探索ノード数
     */
    nodes?: number | null;
    /**
     * Pv
     *
     * 読み筋（PV）
     */
    pv?: Array<string> | null;
};

/**
 * ValidationError
 */
export type ValidationError = {
    /**
     * Location
     */
    loc: Array<string | number>;
    /**
     * Message
     */
    msg: string;
    /**
     * Error Type
     */
    type: string;
};

export type RootGetData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/';
};

export type RootGetResponses = {
    /**
     * Successful Response
     */
    200: unknown;
};

export type HealthCheckHealthGetData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/health';
};

export type HealthCheckHealthGetResponses = {
    /**
     * Successful Response
     */
    200: unknown;
};

export type GetEngineInfoEngineInfoGetData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/engine/info';
};

export type GetEngineInfoEngineInfoGetResponses = {
    /**
     * Successful Response
     */
    200: EngineInfoResponse;
};

export type GetEngineInfoEngineInfoGetResponse = GetEngineInfoEngineInfoGetResponses[keyof GetEngineInfoEngineInfoGetResponses];

export type AnalyzePositionAnalyzePostData = {
    body: AnalysisRequest;
    path?: never;
    query?: never;
    url: '/analyze';
};

export type AnalyzePositionAnalyzePostErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type AnalyzePositionAnalyzePostError = AnalyzePositionAnalyzePostErrors[keyof AnalyzePositionAnalyzePostErrors];

export type AnalyzePositionAnalyzePostResponses = {
    /**
     * Successful Response
     */
    200: AnalysisResponse;
};

export type AnalyzePositionAnalyzePostResponse = AnalyzePositionAnalyzePostResponses[keyof AnalyzePositionAnalyzePostResponses];

export type SearchMateMatePostData = {
    body: MateSearchRequest;
    path?: never;
    query?: never;
    url: '/mate';
};

export type SearchMateMatePostErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type SearchMateMatePostError = SearchMateMatePostErrors[keyof SearchMateMatePostErrors];

export type SearchMateMatePostResponses = {
    /**
     * Successful Response
     */
    200: MateSearchResponse;
};

export type SearchMateMatePostResponse = SearchMateMatePostResponses[keyof SearchMateMatePostResponses];